refactor: 앱 전체의 user 엔티티 import 업데이트
